--- tests/accessibility_tests/test_accessibility.py
@@ -1,7 +1,0 @@
-from playwright.sync_api import Page
-from .common_steps.accessibility import test_accessibility
-import pytest
-
-@pytest.mark.usefixtures("live_server")
-def test_base_accessibility(page: Page, test_accessibility):
-    pass
