--- tests/conftest.py
@@ -1,31 +1,0 @@
-import pytest
-from app import Config
-from app import create_app
-from flask import url_for
-
-class TestConfig(Config):
-    TESTING = True
-    DEBUG = True
-    SERVER_NAME = "localhost"
-    RATELIMIT_ENABLED = False
-    SECRET_KEY = "TEST_KEY"
-
-@pytest.fixture(scope="session")
-def app():
-    app = create_app(TestConfig)
-    yield app
-
-
-@pytest.fixture()
-def client(app):
-    return app.test_client()
-
-
-@pytest.fixture()
-def runner(app):
-    return app.test_cli_runner()
-
-
-@pytest.fixture(scope="function", autouse=True)
-def startup(app, page):
-    page.goto(url_for('main.index', _external=True))
